+++ doc/source/doxygen-docs/changelog.md
@@ -1,6 +1,9 @@
 \page changelog Change Log
 
 # Version 2.8.1: UNRELEASED
+- Changes in libraries:
+  - \ref mrpt_opengl_grp
+    - Expose shadow bias parameters into mrpt::opengl::TLightParameters
 - BUG FIXES:
   - OpenGL shadow rendering: Fix error in calculation of anti peter-panning and shadow acne effects.